move prediction functions into separate and add comments
