+++ front.py
@@ -7,12 +7,10 @@ defines routes for our index.html and instructions.html pages.
 
 import os
 from flask import Flask, flash, request, redirect, render_template
-import decimal
 from werkzeug.utils import secure_filename
-from fastai.basic_train import load_learner
-from fastai.vision import open_image
 from flask_cors import CORS
 import requests
+from makePredictions import *
 
 UPLOAD_FOLDER = os.path.join('static', 'uploads')
 ALLOWED_EXTENSIONS = {'png', 'jpg', 'jpeg', 'gif'}
@@ -99,15 +97,6 @@ def upload_file():
         flash('Please select a file with a .png, .jpg, .jpeg, or .gif extension')
     return render_template("index.html", uploadedImagePath = os.path.join('static', "uploadPH.jpg"))
 
-
-def correct_round(arr, num):
-  newList = []
-  for value in arr:
-    d = decimal.Decimal(str(value))
-    decimal.getcontext().prec=num
-    newList.append(d*1)
-  return newList
-
 ### Google API helper ###
 def get_images(query):
     """ Google API helper. Given a string query, returns a Google image search for that query. """
@@ -121,100 +110,3 @@ def get_images(query):
             results.append(img["link"])
 
     return results
-
-### STYLE ###
-# load the learner
-style_learn = load_learner(path='./models', file='style_unfreeze_300.pkl')
-style_classes = style_learn.data.classes
-style_list = ['Abstract Expressionism', 'Action Painting', 'Analytical Cubism',
-            'Art Nouveau', 'Baroque', 'Color Field Painting',
-            'Contemporary Realism', 'Cubism', 'Early_Renaissance',
-            'Expressionism', 'Fauvism', 'High Renaissance', 'Impressionism',
-            'Mannerism Late Renaissance', 'Minimalism',
-            'Naive Art Primitivism', 'New Realism', 'Northern Renaissance',
-            'Pointillism', 'Pop Art', 'Post Impressionism', 'Realism',
-            'Rococo', 'Romanticism', 'Symbolism', 'Synthetic Cubism', 'Ukiyo e']
-
-# make prediction and load into json
-def predict_style_category(img_file):
-    """ Given an image file, returns the top style prediction. """
-    prediction = style_learn.predict(open_image(img_file))
-    prediction_key = style_classes[prediction[1].item()]
-    return "Predicted Category: " + str(prediction_key)
-
-
-def predict_style_prob(img_file):
-    """ Given an image file, returns the top five style predictions. """
-    prediction = style_learn.predict(open_image(img_file))
-    probs_list = prediction[2].numpy()
-    prob_sorted = sorted(probs_list, key=lambda x: float(x), reverse=True)
-    prob_rounded = correct_round(prob_sorted, 2)
-    percent_dict = {c: str(100*prob_rounded[i]) + "%" for (i, c) in enumerate(style_classes)}
-    percent_list = [str(i).replace(',',':') for i in list(percent_dict.items())]
-
-    top_five = str(percent_list[:5])
-    special_chars = ['[', ']', "'", '"']
-    for i in special_chars:
-      top_five = top_five.replace(i, "")
-
-    return "Top 5 Probabilities: " + top_five
-
-
-### GENRE ###
-# load the learner
-genre_learn = load_learner(path='./models', file='genreLRChanged.pkl')
-genre_classes = genre_learn.data.classes
-
-# make prediction and load into json
-def predict_genre_category(img_file):
-    """ Given an image file, returns the top genre prediction. """
-    prediction = genre_learn.predict(open_image(img_file))
-    prediction_key = genre_classes[prediction[1].item()]
-    return "Predicted Category: " + str(prediction_key)
-
-
-def predict_genre_prob(img_file):
-    """ Given an image file, returns the top five genre predictions. """
-    prediction = genre_learn.predict(open_image(img_file))
-    probs_list = prediction[2].numpy()
-    prob_sorted = sorted(probs_list, key=lambda x: float(x), reverse=True)
-    prob_rounded = correct_round(prob_sorted, 2)
-    percent_dict = {c: str(100*prob_rounded[i]) + "%" for (i, c) in enumerate(genre_classes)}
-    percent_list = [str(i).replace(',',':') for i in list(percent_dict.items())]
-
-    top_five = str(percent_list[:5])
-    special_chars = ['[', ']', "'", '"']
-    for i in special_chars:
-      top_five = top_five.replace(i, "")
-
-    return "Top 5 Probabilities: " + top_five
-
-
-### ARTIST ###
-# load the learner
-artist_learn = load_learner(path='./models', file='artistLR2.pkl')
-artist_classes = artist_learn.data.classes
-
-# make prediction and load into json
-def predict_artist_category(img_file):
-    """ Given an image, returns the top artist prediction. """
-    prediction = artist_learn.predict(open_image(img_file))
-    prediction_key = artist_classes[prediction[1].item()]
-    return "Predicted Category: " + str(prediction_key)
-
-
-def predict_artist_prob(img_file):
-    """ Given an image, returns the top five artist predictions. """
-    prediction = artist_learn.predict(open_image(img_file))
-    probs_list = prediction[2].numpy()
-    prob_sorted = sorted(probs_list, key=lambda x: float(x), reverse=True)
-    prob_rounded = correct_round(prob_sorted, 2)
-    percent_dict = {c: str(100*prob_rounded[i]) + "%" for (i, c) in enumerate(artist_classes)}
-    percent_list = [str(i).replace(',',':') for i in list(percent_dict.items())]
-
-    top_five = str(percent_list[:5])
-    special_chars = ['[', ']', "'", '"']
-    for i in special_chars:
-      top_five = top_five.replace(i, "")
-
-    return "Top 5 Probabilities: " + top_five
